Default to page 1 if offset is too high
Fixed issue with server throwing 500 error when an offset/page combo was requested that didn't exist.  For example if a user requested /?page=500 and we only had 10 pages worth of results, this would result in an infinite loop trying to grab those events.  This fix will default to returning the first page of results.